requirements.txt
flask
python-dotenv
pillow
numpy
tensorflow-cpu==2.13.0
opencv-python-headless
imagehash
itsdangerous
cloudinary
psycopg2-binary
werkzeug
gunicorn